+++ meerschaum/_internal/shell/Shell.py
@@ -235,6 +235,23 @@ def get_shell_intro(with_color: bool = True) -> str:
         **get_config('shell', 'ansi', 'intro', 'rich')
     )
 
+def get_shell_session():
+    """
+    Return the `prompt_toolkit` prompt session.
+    """
+    from meerschaum.config._paths import SHELL_HISTORY_PATH
+    if 'session' in shell_attrs:
+        return shell_attrs['session']
+
+    shell_attrs['session'] = prompt_toolkit_shortcuts.PromptSession(
+        history=prompt_toolkit_history.FileHistory(SHELL_HISTORY_PATH.as_posix()),
+        auto_suggest=ValidAutoSuggest(),
+        completer=ShellCompleter(),
+        complete_while_typing=True,
+        reserve_space_for_menu=False,
+    )
+    return shell_attrs['session']
+
 
 class Shell(cmd.Cmd):
     """
@@ -264,14 +281,7 @@ class Shell(cmd.Cmd):
         except AttributeError:
             pass
 
-        from meerschaum.config._paths import SHELL_HISTORY_PATH
-        shell_attrs['session'] = prompt_toolkit_shortcuts.PromptSession(
-            history=prompt_toolkit_history.FileHistory(SHELL_HISTORY_PATH.as_posix()),
-            auto_suggest=ValidAutoSuggest(),
-            completer=ShellCompleter(),
-            complete_while_typing=True,
-            reserve_space_for_menu=False,
-        )
+        _ = get_shell_session()
 
         super().__init__()